Refactored: Handle state with lifecycle
